Fix a potential bug on iOS 15+ causing no image to be displayed in `FeedImageCell`
The bug is caused by a change in the cell lifecycle behaviour. Now it's possible that a cell will be recreated without calling the `cellForRow` method (if it's already cached). Since we are canceling image loading requests on `didEndDisplayingCell` we must ensure that we reload the appropriate image on `willDisplayCell`.

See https://github.com/essentialdevelopercom/essential-feed-case-study/pull/70 for reference