Ensure all store operations run in the same scheduler
When using `.subscribe(on: scheduler)` we change the scheduler for ALL upstream publishers. This can lead to scenarios, in which, for example, network requests are using the same scheduler we use for our store operations. To avoid this, we replace `.subscribe(on: scheduler)` with `.receive(on: scheduler)` to only affect downstream subscribers.

See https://github.com/essentialdevelopercom/essential-feed-case-study/commit/952ace7603c55aa9ab5f59d417b6d500e9e86fba for reference and further information